requirements.txt
ccxt>=4.1.0
pydantic>=2.5.0
python-telegram-bot>=20.7
numpy>=1.24.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
orjson>=3.9.0
uvloop>=0.19.0
mypy>=1.7.0
ruff>=0.1.0
black>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
psutil>=5.9.0
